flashlight: Add a config option to allow operation in low power modes
The flashlight controller sets up a series of callbacks to prevent
operation in low power modes. This patch allows disabling that behavior,
allowing the user to keep using the flashlight if he so wishes.

Configuring is a compile-time defconfig MTK_FLASHLIGHT_ALLOW_ON_LOW_POWER
which this patch also enables for the porridge and porridgek3 products.